requirements.txt
asgiref==3.3.1
astroid==2.4.2
autopep8==1.6.0
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.1.4
django-coreapi==2.3.0
django-crontab==0.7.1
django-filter==21.1
django-import-export==2.5.0
djangorestframework==3.12.2
environs==9.5.0
et-xmlfile==1.0.1
idna==2.10
isort==5.6.4
itypes==1.2.0
Jinja2==2.11.3
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
marshmallow==3.15.0
mccabe==0.6.1
odfpy==1.4.1
openpyxl==3.0.7
packaging==21.3
Pillow==8.0.1
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pylint==2.6.0
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2020.4
PyYAML==5.4.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
toml==0.10.2
uritemplate==3.0.1
urllib3==1.26.2
wrapt==1.12.1
xlrd==1.0.0
xlwt==1.3.0
